requirements.txt
arrow==1.2.2
attrs==22.1.0
beautifulsoup4==4.11.1
black==22.8.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.3
flake8==5.0.4
Flask==2.2.5
gunicorn>=20.1.0
html5lib>=1.1
ics==0.7.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml>=4.9.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.3
pathspec==0.10.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.31.0
six==1.16.0
soupsieve==2.3.2.post1
TatSu==5.8.3
tomli==2.0.1
urllib3==1.26.18
Werkzeug==3.0.1
pyyaml>=6.0
canvasapi>=3.1.0
pandas>=1.4.3
html5lib
canvas-crawler

